Removed phone-view folder and moved remaining quiz view into participantQuizView
