+++ client/src/pages/participantQuizView/components/HasAnsweredView.tsx
