+++ client/src/pages/participantQuizView/components/LobbyView.tsx
