+++ client/src/pages/participantQuizView/components/QuizEndedView.tsx
